requirements.txt
playwright
pandas
openai
python-dotenv